Move free-version related methods to a separate file
