+++ cmd/xl-storage-format-v2.go
@@ -102,6 +102,11 @@ func isXL2V1Format(buf []byte) bool {
 //
 // The most recently updated element in the array is considered the latest version.
 
+// In addition to these we have a special kind called free-version. This is represented
+// using a delete-marker and MetaSys entries. It's used to track tiered content of a
+// deleted/overwritten version. This version is visible _only_to the scanner routine, for subsequent deletion.
+// This kind of tracking is necessary since a version's tiered content is deleted asynchronously.
+
 // Backend directory tree structure:
 // disk1/
 // └── bucket
@@ -227,15 +232,6 @@ func (j xlMetaV2Version) Valid() bool {
 	case DeleteType:
 		return j.DeleteMarker != nil &&
 			j.DeleteMarker.ModTime > 0
-	}
-	return false
-}
-
-// FreeVersion returns true if j represents a free-version, false otherwise.
-func (j xlMetaV2Version) FreeVersion() bool {
-	switch j.Type {
-	case DeleteType:
-		return j.DeleteMarker.FreeVersion()
 	}
 	return false
 }
@@ -938,37 +934,6 @@ func (z *xlMetaV2) AddVersion(fi FileInfo) error {
 	return nil
 }
 
-// AddFreeVersion adds a free-version if needed for fi.VersionID version.
-// Free-version will be added if fi.VersionID has transitioned.
-func (z *xlMetaV2) AddFreeVersion(fi FileInfo) error {
-	var uv uuid.UUID
-	var err error
-	switch fi.VersionID {
-	case "", nullVersionID:
-	default:
-		uv, err = uuid.Parse(fi.VersionID)
-		if err != nil {
-			return err
-		}
-	}
-
-	for _, version := range z.Versions {
-		switch version.Type {
-		case ObjectType:
-			if version.ObjectV2.VersionID == uv {
-				// if uv has tiered content we add a
-				// free-version to track it for asynchronous
-				// deletion via scanner.
-				if freeVersion, toFree := version.ObjectV2.InitFreeVersion(fi); toFree {
-					z.Versions = append(z.Versions, freeVersion)
-				}
-				return nil
-			}
-		}
-	}
-	return nil
-}
-
 func (j xlMetaV2DeleteMarker) ToFileInfo(volume, path string) (FileInfo, error) {
 	versionID := ""
 	var uv uuid.UUID
@@ -1131,50 +1096,6 @@ func (z *xlMetaV2) SharedDataDirCount(versionID [16]byte, dataDir [16]byte) int 
 		}
 	}
 	return sameDataDirCount
-}
-
-// FreeVersion returns ture if j represents a free-version, false otherwise.
-func (j xlMetaV2DeleteMarker) FreeVersion() bool {
-	_, ok := j.MetaSys[ReservedMetadataPrefixLower+freeVersion]
-	return ok
-}
-
-// Free-version is a special version meant to track tiered content of a
-// deleted/overwritten version. This version is visible only to the scanner
-// routine, for subsequent deletion. This kind of tracking is necessary since
-// the tiered content of an object version is deleted asynchronously.
-
-const freeVersion = "free-version"
-
-// InitFreeVersion creates a free-version to track the tiered-content of j. If j has
-// no tiered content, it returns false.
-func (j xlMetaV2Object) InitFreeVersion(fi FileInfo) (xlMetaV2Version, bool) {
-	if status, ok := j.MetaSys[ReservedMetadataPrefixLower+TransitionStatus]; ok && bytes.Equal(status, []byte(lifecycle.TransitionComplete)) {
-		vID, err := uuid.Parse(fi.TierFreeVersionID())
-		if err != nil {
-			panic(fmt.Errorf("Invalid Tier Object delete marker versionId %s %v", fi.TierFreeVersionID(), err))
-		}
-		freeEntry := xlMetaV2Version{Type: DeleteType}
-		freeEntry.DeleteMarker = &xlMetaV2DeleteMarker{
-			VersionID: vID,
-			ModTime:   j.ModTime, // fi.ModTime may be empty
-			MetaSys:   make(map[string][]byte),
-		}
-
-		freeEntry.DeleteMarker.MetaSys[ReservedMetadataPrefixLower+freeVersion] = []byte{}
-		tierKey := ReservedMetadataPrefixLower + TransitionTier
-		tierObjKey := ReservedMetadataPrefixLower + TransitionedObjectName
-		tierObjVIDKey := ReservedMetadataPrefixLower + TransitionedVersionID
-
-		for k, v := range j.MetaSys {
-			switch k {
-			case tierKey, tierObjKey, tierObjVIDKey:
-				freeEntry.DeleteMarker.MetaSys[k] = v
-			}
-		}
-		return freeEntry, true
-	}
-	return xlMetaV2Version{}, false
 }
 
 // DeleteVersion deletes the version specified by version id.
